% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_fns.R
\name{compute_isco_share}
\alias{compute_isco_share}
\title{Compute ISCO-Level Wage Bill Shares}
\usage{
compute_isco_share(...)
}
\arguments{
\item{...}{Arguments passed to \code{compute_wagebill_byoccup()}.}
}
\value{
A tibble with columns:
\describe{
  \item{occupation_iscocode}{ISCO code of the occupation group.}
  \item{occupation_isconame}{Name of the occupation group.}
  \item{year}{Year of observation.}
  \item{indicator}{Name of the wage variable.}
  \item{isco_share}{Occupation’s share of the total wage bill.}
}
}
\description{
Calculates each occupation's share of the total wage bill within a given year and indicator.
}
\details{
This function first computes the wage bill by occupation using \code{compute_wagebill_byoccup()},
then calculates the proportional share of each occupation’s wage bill in the total for the same
year and wage indicator.
}
