% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi}
\alias{cpi}
\title{Consumer Price Index}
\format{
A data frame with 9078 rows and 3 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code}
  \item{\code{year}}{character Year}
  \item{\code{cpi}}{character Consumer Price Index (2010 = 100)}
}
}
\source{
<https://data360.worldbank.org/en/int/indicator/WB_WDI_FP_CPI_TOTL>
}
\usage{
cpi
}
\description{
Index of the prices of consumption goods and services, as compared to a certain reference period (2010=100).
}
\keyword{datasets}
