% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enterprise_surveys}
\alias{enterprise_surveys}
\title{World Bank Enterprise Surveys}
\format{
A data frame with 413 rows and 3 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code.}
  \item{\code{year}}{character Year.}
  \item{\code{workforce_constraint}}{character Proportion of establishments identifying an inadequately educated workforce as a major or severe constraint to the current operations of the establishment.}
}
}
\source{
World Bank
\url{https://data360.worldbank.org/en/int/indicator/WB_ES_T_WK10}
}
\usage{
enterprise_surveys
}
\description{
The Formal Sector World Bank Enterprise Surveys, generally known by the shortened term, World Bank Enterprise Surveys (WBES), are the main product of the Enterprise Surveys team. They are nationally representative firm-level surveys with top managers and owners of businesses in over 160 economies, reaching 180 in upcoming years, that provide insight into many business environment topics such as access to finance, corruption, infrastructure, and performance, among others. The comprehensive data and analytical reports enable easy comparisons across economies and time. The information collected through these surveys is publicly available at the economy and firm level. Explore the indicators by economy or topic, browse the surveys, or access the granular data on the WBES data portal.
}
\keyword{datasets}
