% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data360api.R
\name{get_data360_api}
\alias{get_data360_api}
\title{Fetch World Bank Data360 CPI}
\usage{
get_data360_api(dataset_id, indicator_id)
}
\arguments{
\item{dataset_id}{Unique identifier for the database.}

\item{indicator_id}{Indicator ID.}
}
\value{
A tibble containing data, including country codes, country names,
        years, and values.
}
\description{
This function retrieves data from the Data360 API.
}
\examples{
\dontrun{
  cpi_data <- get_data360_api("WB_WDI", "WB_WDI_FP_CPI_TOTL")
  head(cpi_data)
}

}
