% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggplot_point_line}
\alias{ggplot_point_line}
\title{Plot grouped line chart with points}
\usage{
ggplot_point_line(data, x, y, group = NULL, ...)
}
\arguments{
\item{data}{A data frame or tibble containing the variables to plot.}

\item{x}{Column to be mapped to the x-axis.}

\item{y}{Column to be mapped to the y-axis.}

\item{group}{Column specifying the grouping variable.
This will be mapped to color.}
}
\value{
A \code{ggplot} object representing the grouped line chart.
}
\description{
This function creates a line plot with points, showing how a variable
(y) evolves along another variable (x)
}
\examples{
\dontrun{
ggplot_point_line(
  data,
  x = year,
  y = median_base_salary_ppp,
  group = paygrade
)
}

}
