% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_fns.R
\name{harmonize_col_class}
\alias{harmonize_col_class}
\title{Harmonize Column Classes Across Existing Columns in a List of Data Frames}
\usage{
harmonize_col_class(df_list)
}
\arguments{
\item{df_list}{A list of data frames to harmonize.}
}
\value{
A list of data frames with shared columns coerced to the same class if needed.
}
\description{
This function takes a list of data frames and ensures that any columns shared
across two or more data frames have consistent classes. If a column has
different classes across data frames, it is coerced to character in those
data frames. Columns that are missing from a data frame are left untouched
(i.e., no new columns are added).
}
\examples{
df1 <- data.frame(a = 1:3, b = letters[1:3])
df2 <- data.frame(b = factor(letters[4:6]), c = c("x", "y", "z"))
result <- harmonize_col_class(list(df1, df2))
lapply(result, str)

}
