% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_fns.R
\name{merge_wrapper}
\alias{merge_wrapper}
\title{Merge Wrapper Function}
\usage{
merge_wrapper(...)
}
\arguments{
\item{...}{Arguments passed to [base::merge()], such as \code{x}, \code{y},
\code{by}, \code{by.x}, \code{by.y}, etc.}
}
\value{
A data frame containing the merged result with all rows preserved.
}
\description{
A simple wrapper around [base::merge()] that performs a full outer join
(i.e., keeps all rows from both data frames).
}
\examples{
df1 <- data.frame(id = 1:3, value1 = c("A", "B", "C"))
df2 <- data.frame(id = 2:4, value2 = c("X", "Y", "Z"))
merge_wrapper(df1, df2, by = "id")

}
