% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppp}
\alias{ppp}
\title{PPP Conversion Factor}
\format{
A data frame with 6172 rows and 3 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code}
  \item{\code{year}}{character Year}
  \item{\code{ppp}}{character PPP conversion factor, private consumption (LCU per international $)}
}
}
\source{
<https://data360.worldbank.org/en/int/indicator/WB_WDI_PA_NUS_PRVT_PP>
}
\usage{
ppp
}
\description{
The purchasing power parity (PPP) conversion factor is a currency conversion factor and a spatial price deflator. They convert different currencies to a common currency and, in the process of conversion, equalize their purchasing power by eliminating the differences in price levels between countries, thereby allowing volume or output comparisons of gross domestic product (GDP) and its expenditure components. This conversion factor is for household final consumption expenditure and the base currency is the US dollar.
}
\keyword{datasets}
