% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitychecks.R
\name{qualitycheck_contractmod}
\alias{qualitycheck_contractmod}
\title{Quality Checks for Harmonized HRMIS Contract Data}
\usage{
qualitycheck_contractmod(contract_tbl, required_vars = NULL)
}
\arguments{
\item{contract_tbl}{A harmonized HRMIS contract tibble (e.g. contract_alagoas_tbl)}

\item{required_vars}{A character vector of required variable names (from dictionary)}
}
\value{
A pointblank agent with a battery of test results
}
\description{
Quality Checks for Harmonized HRMIS Contract Data
}
