% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitychecks.R
\name{qualitycheck_orgmod}
\alias{qualitycheck_orgmod}
\title{Quality Check for Harmonized Organization Module}
\usage{
qualitycheck_orgmod(org_tbl)
}
\arguments{
\item{org_tbl}{A data.frame or tibble containing the harmonized organization module}
}
\value{
A pointblank agent object
}
\description{
Quality Check for Harmonized Organization Module
}
