% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social_sustainability}
\alias{social_sustainability}
\title{Social Sustainability Global Database}
\format{
A data frame with 195 rows and 3 variables:
\describe{
  \item{\code{country_code}}{character World Bank country code.}
  \item{\code{year}}{numeric Year.}
  \item{\code{confidence_in_gov}}{numeric Percentage of population with confidence in government.}
}
}
\source{
World Bank
\url{https://datacatalog.worldbank.org/int/search/dataset/0061880/Social-Sustainability-Global-Database-}
}
\usage{
social_sustainability
}
\description{
The Social Sustainability global database and its visualization dashboard <https://public.tableau.com/app/profile/social.sustainability.and.inclusion.world.bank/viz/SocialSustainabilityGlobalDashboard2_0/Historia1?publish=yes/> are global public goods produced by the Social Development Global Practice of The World Bank Group. They feature leading indicators of inclusion, resilience, social cohesion, and process legitimacy, for 222 countries, disaggregated by population group and analyzed spatially and over time. In addition, the dashboard allows the user to overlay the indicators in the geospatial platform of the World Bank Group.
}
\keyword{datasets}
