% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization_fns.R
\name{vectorize_gt}
\alias{vectorize_gt}
\title{A function to translate a vector to language of choice}
\usage{
vectorize_gt(vector, target_language = "en", source_language)
}
\arguments{
\item{vector}{object containing the list of characters to be translated}

\item{target_language}{two letter ISO language name for `vector` to be translated
into}

\item{source_language}{two letter ISO language name for `vector` to be translated
from}
}
\description{
This function vectorizes the `polyglotr::google_translate()` function to support
translation from `source_language` to any `target_language` of choice.
}
