% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_fns.R
\name{compute_share}
\alias{compute_share}
\title{Compute ratio (share) indicators relative to macro data}
\usage{
compute_share(
  data,
  macro_data = macro_indicators,
  cols = list(vars = NULL, groups = c("country_code", "year")),
  macro_cols,
  fns = list(default = c("sum"), user = NULL),
  output = c("long", "wide")
)
}
\arguments{
\item{data}{A data frame containing the microdata.}

\item{macro_data}{A data frame containing macro-level indicators.}

\item{cols}{A list of 2 elements defining variables (`vars`) and groups (`groups`)
for `compute_summary()` (see `?compute_summary`).}

\item{macro_cols}{Character vector of macro indicator columns to use.}

\item{fns}{A list of summary functions (default and/or user-defined).}

\item{output}{Either `"long"` or `"wide"` (default `"long"`).}
}
\value{
A tibble containing ratio indicators (shares) in long or wide format.
}
\description{
Computes ratios of summarized values from a dataset to corresponding macro-level indicators
(e.g., wage bill as a share of GDP or population). Uses `compute_summary()` internally.
}
\details{
Ensure the `groups` argument match with `colnames` within the
`macro_data` for appropriate comparisons between results of summarizing
`vars` (within cols argument) and `macro_data`. The current default is
set for the HRM harmonization which assumes shares of `vars` in the
`macro_data` which is merged by `country_code` and `year` for appropriate
comparisons.
}
