% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitychecks.R
\name{qualitycheck_worker}
\alias{qualitycheck_worker}
\title{Quality Check for Harmonized Worker Module}
\usage{
qualitycheck_worker(worker_tbl)
}
\arguments{
\item{worker_tbl}{A data.frame or tibble containing the harmonized worker module}
}
\value{
A pointblank agent object
}
\description{
Quality Check for Harmonized Worker Module
}
